requirements.txt
boto3==1.40.59
botocore==1.40.59
jmespath==1.0.1
pillow==12.0.0
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
